requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
web3==5.30.0
python-dotenv==0.21.0
pytest==7.1.3